lib/api/requirements.txt
Flask==3.1.0
numpy==1.26.4
markupsafe==2.0.1
flask-cors==5.0.0
transformers==4.47.0
pillow==11.0.0
timm==1.0.10
einops==0.8.0
